+++ compose/src/commonMain/kotlin/ComposeNavigationExtensions.kt
@@ -1,7 +1,7 @@
 package dev.inmo.navigation.compose
 
 import androidx.compose.runtime.*
-import dev.inmo.micro_utils.coroutines.subscribeSafelyWithoutExceptions
+import dev.inmo.micro_utils.coroutines.subscribeLoggingDropExceptions
 import dev.inmo.navigation.core.NavigationChain
 import dev.inmo.navigation.core.NavigationChainId
 import dev.inmo.navigation.core.NavigationNode
@@ -10,25 +10,35 @@ import dev.inmo.navigation.core.onDestroyFlow
 import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.flow.take
 
+@Composable
+internal fun <Base> DrawNode(node: NavigationNode<out Base, Base>) {
+    when {
+        node is ComposeNode -> {
+            val drawState = node.drawerState.collectAsState()
+            drawState.value ?.invoke()
+        }
+
+        else -> {
+            SubchainsHandling<Base>()
+        }
+    }
+}
 /**
- * Collecting [NavigationChain.stackFlow] and always [Draw] only the latest element in [NavigationChain.stackFlow]
+ * Collecting [NavigationChain.stackFlow] and always [DrawChain] only the latest element in [NavigationChain.stackFlow]
  */
 @Composable
-internal fun <Base> NavigationChain<Base>.DrawStackNodes() {
-    val stack = stackFlow.collectAsState()
-    key(stack.value.lastOrNull()) {
-        val latestNode = stack.value.lastOrNull()
+internal fun <Base> DrawStackNodes() {
+    val chain = getChainFromLocalProvider<Base>() ?: return
+    val stack = chain.stackFlow.collectAsState()
+    val latestNode = remember(stack.value.lastOrNull()) {
+        stack.value.lastOrNull()
+    }
+    val provider = InternalLocalComposeInjectedChainsAndNodesProvider.current
+    key(latestNode, latestNode ?.hashCode()) {
         if (latestNode != null) {
             doWithNodeInLocalProvider(latestNode) {
-                when {
-                    latestNode is ComposeNode -> {
-                        val drawState = latestNode.drawerState.collectAsState()
-                        drawState.value ?.invoke()
-                    }
-
-                    else -> {
-                        latestNode.SubchainsHandling()
-                    }
+                if (latestNode !in provider.nodes) {
+                    DrawNode(latestNode)
                 }
             }
         }
@@ -36,15 +46,22 @@ internal fun <Base> NavigationChain<Base>.DrawStackNodes() {
 }
 
 /**
- * Calls [Draw] on each [NavigationChain] in [NavigationNode.subchainsFlow] of [this]
+ * Calls [DrawChain] on each [NavigationChain] in [NavigationNode.subchainsFlow] of [this]
  */
 @Composable
-internal fun <Base> NavigationNode<out Base, Base>.SubchainsHandling(filter: (NavigationChain<Base>) -> Boolean = { true }) {
-    val subchainsState = subchainsFlow.collectAsState()
+internal fun <Base> SubchainsHandling(filter: (NavigationChain<Base>) -> Boolean = { true }) {
+    val node = getNodeFromLocalProvider<Base>() ?: return
+    val subchainsState = node.subchainsFlow.collectAsState()
     val rawSubchains = subchainsState.value
     val filteredSubchains = rawSubchains.filter(filter)
+
+    val provider = InternalLocalComposeInjectedChainsAndNodesProvider.current
     filteredSubchains.forEach {
-        it.Draw()
+        if (provider.chains.contains(it) == false) {
+            doWithChainInLocalProvider(it) {
+                DrawChain<Base>()
+            }
+        }
     }
 }
 
@@ -55,46 +72,27 @@ internal fun <Base> NavigationNode<out Base, Base>.SubchainsHandling(filter: (Na
  * @param beforeNodes Will be called **before** [DrawStackNodes] will be called
  */
 @Composable
-internal fun <Base> NavigationChain<Base>.Draw(
+internal fun <Base> DrawChain(
     onDismiss: (suspend NavigationChain<Base>.() -> Unit)? = null,
     beforeNodes: @Composable NavigationChain<Base>.() -> Unit = {  }
 ) {
-    key(onDismiss, parentNode) {
+    val chain = getChainFromLocalProvider<Base>() ?: return
+    key(onDismiss, chain.parentNode) {
         onDismiss ?.let {
-            parentNode ?.let { parentNode ->
+            chain.parentNode ?.let { parentNode ->
                 val scope = rememberCoroutineScope()
 
                 remember {
-                    parentNode.onChainRemovedFlow.filter { it.any { it.value === this@Draw } }.subscribeSafelyWithoutExceptions(scope) {
-                        onDismiss(this)
+                    parentNode.onChainRemovedFlow.filter { it.any { it.value === chain } }.subscribeLoggingDropExceptions(scope) {
+                        onDismiss(chain)
                     }
                 }
             }
         }
     }
 
-    doWithChainInLocalProvider(this) {
-        beforeNodes()
-        DrawStackNodes()
-    }
-}
-
-/**
- * Using [getNodesFactoryFromLocalProvider] to get navigation nodes factory, creating new subchain with
- * [NavigationNode.createEmptySubChain] if [this] [NavigationNode] is not null, and creating new one without parent node
- * with [NavigationChain] constructor. After chain created it will call [Draw] on it
- */
-@Composable
-internal fun <Base> NavigationNode<*, Base>?.SubChain(
-    onDismiss: (suspend NavigationChain<Base>.() -> Unit)? = null,
-    id: NavigationChainId? = null,
-    beforeNodes: @Composable NavigationChain<Base>.() -> Unit
-) {
-    val factory = getNodesFactoryFromLocalProvider<Base>()
-    val chain = remember(this, factory) {
-        this ?.createEmptySubChain(id = id) ?: NavigationChain<Base>(parentNode = null, nodeFactory = factory, id = id)
-    }
-    chain.Draw(onDismiss, beforeNodes)
+    chain.beforeNodes()
+    DrawStackNodes<Base>()
 }
 
 /**
@@ -106,14 +104,32 @@ fun <Base> InjectNavigationChain(
     id: NavigationChainId? = null,
     beforeNodes: @Composable NavigationChain<Base>.() -> Unit
 ) {
-    getNodeFromLocalProvider<Base>().SubChain(onDismiss = onDismiss, id = id, beforeNodes = beforeNodes)
+    val rootNode = getNodeFromLocalProvider<Base>()
+    val factory = getNodesFactoryFromLocalProvider<Base>()
+    val chain = remember(rootNode, factory) {
+        if (rootNode == null) {
+            NavigationChain<Base>(parentNode = null, nodeFactory = factory, id = id)
+        } else {
+            rootNode.createEmptySubChain(id = id)
+        }
+    }
+    val provider = InternalLocalComposeInjectedChainsAndNodesProvider.current
+    remember(chain) { provider.chains.add(chain) }
+    DisposableEffect(provider, chain) {
+        onDispose {
+            provider.chains.remove(chain)
+        }
+    }
+    doWithChainInLocalProvider(chain) {
+        DrawChain(onDismiss, beforeNodes)
+    }
 }
 
 /**
  * Just calls [InjectNavigationChain]
  */
 @Composable
-@Deprecated("Renamed", ReplaceWith("InjectNavigationChain(onDismiss, block)", "dev.inmo.navigation.compose.InjectNavigationChain"))
+@Deprecated("Renamed", ReplaceWith("InjectNavigationChain(onDismiss, id, block)", "dev.inmo.navigation.compose.InjectNavigationChain"))
 fun <Base> SubChain(
     onDismiss: (suspend NavigationChain<Base>.() -> Unit)? = null,
     id: NavigationChainId? = null,
@@ -130,11 +146,11 @@ fun <Base> SubChain(
 @Composable
 internal fun <Base> NavigationNode<*, Base>.Use(
     onDismiss: (suspend NavigationNode<*, Base>.() -> Unit)? = null,
-    actionInContext: @Composable() (() -> Unit)? = null,
+    actionInContext: @Composable() (NavigationNode<out Base, Base>.() -> Unit)? = null,
 ) {
     key(this, actionInContext) {
         doWithNodeInLocalProvider(this) {
-            actionInContext ?.invoke()
+            actionInContext ?.invoke(this)
         }
     }
 
@@ -142,7 +158,7 @@ internal fun <Base> NavigationNode<*, Base>.Use(
         onDismiss ?.let { onDismiss ->
             val scope = rememberCoroutineScope()
 
-            onDestroyFlow.take(1).subscribeSafelyWithoutExceptions(scope) {
+            onDestroyFlow.take(1).subscribeLoggingDropExceptions(scope) {
                 onDismiss(this)
             }
         }
@@ -154,48 +170,68 @@ internal fun <Base> NavigationNode<*, Base>.Use(
  * [onDismiss] and [additionalCodeInNodeContext] to it
  */
 @Composable
-internal fun <Base> NavigationChain<Base>.NodeInStack(
+internal fun <Base> PushAndDrawNodeInStack(
     config: Base,
     onDismiss: (suspend NavigationNode<*, Base>.() -> Unit)? = null,
-    additionalCodeInNodeContext: @Composable() (() -> Unit)? = null,
+    additionalCodeInNodeContext: @Composable() (NavigationNode<out Base, Base>.() -> Unit)? = null,
 ) {
-    val node: NavigationNode<out Base, Base>? = remember {
-        push(config)
+    val chain = getChainFromLocalProvider<Base>() ?: return
+    val node: MutableState<NavigationNode<out Base, Base>?> = remember { mutableStateOf(null) }
+    LaunchedEffect(config) {
+        val currentValue = node.value
+        node.value = chain.push(config)
+        if (currentValue != null) {
+            chain.drop(currentValue)
+        }
     }
-    node ?.Use(onDismiss, additionalCodeInNodeContext)
+    val provider = InternalLocalComposeInjectedChainsAndNodesProvider.current
+    node.value ?.let {
+        remember { provider.nodes.add(it) }
+        DisposableEffect(provider, it) {
+            onDispose {
+                provider.nodes.remove(it)
+            }
+        }
+    }
+    node.value ?.Use(onDismiss, additionalCodeInNodeContext)
+
+    node.value ?.let {
+        DrawNode(it)
+    }
 }
 
 /**
- * Trying to get current [NavigationChain] using [getChainFromLocalProvider] and calls [NodeInStack] with
+ * Trying to get current [NavigationChain] using [getChainFromLocalProvider] and calls [PushAndDrawNodeInStack] with
  * passing both [config], [onDismiss] and [additionalCodeInNodeContext]
  *
- * If [NavigationChain] is absent in context, will create new one with [InjectNavigationChain] and pass calling of [NodeInStack]
- * as [NodeInStack]
+ * If [NavigationChain] is absent in context, will create new one with [InjectNavigationChain] and pass calling of [PushAndDrawNodeInStack]
+ * as [PushAndDrawNodeInStack]
  */
 @Composable
 fun <Base> InjectNavigationNode(
     config: Base,
     onDismiss: (suspend NavigationNode<*, Base>.() -> Unit)? = null,
-    additionalCodeInNodeContext: @Composable() (() -> Unit)? = null,
+    additionalCodeInNodeContext: @Composable() (NavigationNode<out Base, Base>.() -> Unit)? = null,
 ) {
-    val chain = getChainFromLocalProvider<Base>() ?: run {
+    val existsChain = getChainFromLocalProvider<Base>()
+    if (existsChain == null) {
         InjectNavigationChain<Base> {
-            NodeInStack(config, onDismiss, additionalCodeInNodeContext)
+            PushAndDrawNodeInStack(config, onDismiss, additionalCodeInNodeContext)
         }
-        return@InjectNavigationNode
+    } else {
+        PushAndDrawNodeInStack(config, onDismiss, additionalCodeInNodeContext)
     }
-    chain.NodeInStack(config, onDismiss, additionalCodeInNodeContext)
 }
 
 /**
  * Just calling [InjectNavigationNode] with passing arguments as is
  */
 @Composable
-@Deprecated("Renamed", ReplaceWith("InjectNavigationNode(config, onDismiss)", "dev.inmo.navigation.compose.InjectNavigationNode"))
+@Deprecated("Renamed", ReplaceWith("InjectNavigationNode(config, onDismiss, additionalCodeInNodeContext)", "dev.inmo.navigation.compose.InjectNavigationNode"))
 fun <Base> NavigationSubNode(
     config: Base,
     onDismiss: (suspend NavigationNode<*, Base>.() -> Unit)? = null,
-    additionalCodeInNodeContext: @Composable() (() -> Unit)? = null,
+    additionalCodeInNodeContext: @Composable() (NavigationNode<out Base, Base>.() -> Unit)? = null,
 ) {
     InjectNavigationNode(config, onDismiss, additionalCodeInNodeContext)
 }